+++ src/os2datascanner/projects/report/tests/test_handle_clean_message.py
@@ -2,7 +2,8 @@ import json
 
 from django.test import TestCase
 
-from ..reportapp.management.commands.event_collector import handle_clean_message
+from ..reportapp.management.commands.event_collector import (
+    handle_clean_account_message, handle_clean_problem_message)
 from ..reportapp.utils import create_alias_and_match_relations
 from ..organizations.models import Account, Alias, Organization, AliasType
 from ..reportapp.models.documentreport import DocumentReport
@@ -198,7 +199,7 @@ raw_matches_json_matched = json.loads('''
 ''')
 
 
-class HandleCleanMessageTest(TestCase):
+class HandleCleanAccountMessageTest(TestCase):
 
     def setUp(self):
         org = Organization.objects.create(name="TestOrg")
@@ -262,7 +263,7 @@ class HandleCleanMessageTest(TestCase):
             },
             "type": "clean_document_reports"}
 
-        handle_clean_message(message)
+        handle_clean_account_message(message)
 
         self.assertEqual(DocumentReport.objects.count(), 30)
         self.assertEqual(
@@ -287,7 +288,7 @@ class HandleCleanMessageTest(TestCase):
             "scanners_accounts_dict": {},
             "type": "clean_document_reports"}
 
-        handle_clean_message(message)
+        handle_clean_account_message(message)
         self.assertEqual(DocumentReport.objects.count(), 40)
         self.assertEqual(
             DocumentReport.objects.filter(
@@ -311,7 +312,7 @@ class HandleCleanMessageTest(TestCase):
             },
             "type": "clean_document_reports"}
 
-        handle_clean_message(message)
+        handle_clean_account_message(message)
 
         self.assertEqual(DocumentReport.objects.count(), 20)
         self.assertEqual(
@@ -352,7 +353,7 @@ class HandleCleanMessageTest(TestCase):
             },
             "type": "clean_document_reports"}
 
-        handle_clean_message(message)
+        handle_clean_account_message(message)
 
         self.assertEqual(DocumentReport.objects.count(), 20)
         self.assertEqual(
@@ -388,7 +389,7 @@ class HandleCleanMessageTest(TestCase):
             },
             "type": "clean_document_reports"}
 
-        handle_clean_message(message)
+        handle_clean_account_message(message)
 
         self.assertEqual(
             DocumentReport.objects.filter(
@@ -410,3 +411,80 @@ class HandleCleanMessageTest(TestCase):
                 alias_relation__account=self.egon,
                 scanner_job_pk=2).count(),
             0)
+
+
+class HandleCleanProblemMessageTest(TestCase):
+    def setUp(self):
+
+        # Create ten matched reports
+        for i in range(10):
+            DocumentReport.objects.create(
+                name=f"Matched Report-{i}",
+                owner="bøffen",
+                scanner_job_pk=1,
+                path=f"report-{i}-1m-bøffen",
+                raw_matches=raw_matches_json_matched)
+
+        # Create ten unmatched reports
+        for i in range(10):
+            DocumentReport.objects.create(
+                name=f"Unmatched Report-{i}",
+                owner="bøffen",
+                scanner_job_pk=1,
+                path=f"report-{i}-1u-bøffen")
+
+        # Create ten unmatched reports from second scannerjob
+        for i in range(10):
+            DocumentReport.objects.create(
+                name=f"(2) Unmatched Report-{i}",
+                owner="bøffen",
+                scanner_job_pk=2,
+                path=f"report-{i}-2u-bøffen")
+
+    def test_cleaning_problems_one_scanner(self):
+        """Cleaning problems for one scanner should not touch problems from
+        other scanners."""
+        message = {
+            "scanners": [1],
+            "type": "clean_problem_reports"
+        }
+        handle_clean_problem_message(message)
+
+        # Make sure only the ten unmatched reports were deleted
+        self.assertEqual(DocumentReport.objects.all().count(), 20,
+                         "Expected to find 20 document reports after cleaning")
+        self.assertEqual(
+            DocumentReport.objects.filter(
+                scanner_job_pk=1,
+                number_of_matches=0).count(),
+            0,
+            "Not all problem messages related to scanner were deleted")
+
+        # Make sure reports from other scanner are untouched
+        self.assertEqual(DocumentReport.objects.filter(scanner_job_pk=2).count(),
+                         10, "Some reports from unrelated scanner were deleted")
+
+    def test_cleaning_problems_for_two_scanners(self):
+        """Cleaning problems for multiple scanners should remove problems
+        from both."""
+        message = {
+            "scanners": [1, 2],
+            "type": "clean_problem_reports"
+        }
+        handle_clean_problem_message(message)
+
+        # Make sure only problem reports were deleted
+        self.assertEqual(DocumentReport.objects.all().count(), 10,
+                         "Expected to find 10 document reports after cleaning")
+        self.assertEqual(
+            DocumentReport.objects.filter(
+                scanner_job_pk=1,
+                number_of_matches=0).count(),
+            0,
+            "Not all problem messages related to scanner 1 were deleted")
+        self.assertEqual(
+            DocumentReport.objects.filter(
+                scanner_job_pk=2,
+                number_of_matches=0).count(),
+            0,
+            "Not all problem messages related to scanner 2 were deleted")