+++ CHANGELOG.md
@@ -16,9 +16,11 @@
 - DocumentReports without matches are now deleted when the corresponding
   source no longer exists.
 
-
 - It is now possible to interrupt an organisational import job while, or
   before, it's running.
+
+- When a scanner is deleted, the DocumentReports with no matches related to
+  that scanner are also deleted.
 
 ### Bugfixes